requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
email-validator==2.2.0
mysqlclient==2.2.4
SQLAlchemy==2.0.35
python-dotenv==1.0.1
WTForms==3.1.2
cryptography==43.0.1
black==24.8.0
flake8==7.1.1
